Ensure new track plays from start after replacement (#95)
Added a seek operation to reset the playback position when replacing a track in ExoPlayer. Previously, after replacing a track while in Player.STATE_ENDED, the player would maintain its position at the end of the old track. By seeking to the start of the newly added track, we ensure that playback begins from the beginning of the new media item.